blackfridayapi/requirements.txt
cachetools==5.3.3
charset-normalizer==3.3.2
Django==2.2.24
djangorestframework==3.11.0
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-aiplatform==1.34.0
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.48.2
idna==3.7
importlib-metadata==6.7.0
packaging==24.0
proto-plus==1.23.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
wincertstore==0.2
zipp==3.15.0
